recommender/requirements.txt
# Recommender System Requirements
pandas>=2.0.0
numpy>=1.24.0
google-cloud-bigquery>=3.11.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
db-dtypes>=1.1.1

# API REST
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Vertex AI
google-cloud-aiplatform>=1.38.0